fix: change the style of the prev/next question btn on the game page
