+++ apps/client/components/main/Tips.vue
@@ -18,24 +18,6 @@
       </button>
       <span class="ml-2">{{ toggleTipText }}</span>
     </div>
-    <div class="mb-4">
-      <button
-        class="mr-1 tip-btn"
-        @click="goToPreviousQuestion"
-      >
-        ⌃ {{ shortcutKeys.previous }}
-      </button>
-      <span class="ml-2">上一题</span>
-    </div>
-    <div class="mb-4">
-      <button
-        class="mr-1 tip-btn"
-        @click="goToNextQuestion"
-      >
-        ⌃ {{ shortcutKeys.skip }}
-      </button>
-      <span class="ml-2">下一题</span>
-    </div>
     <div>
       <button class="tip-btn">Space</button>
       <span class="ml-2">{{ spaceTipText }} </span>
@@ -50,20 +32,11 @@ import { useCurrentStatementEnglishSound } from "~/composables/main/englishSound
 import { useGameMode } from "~/composables/main/game";
 import { useSummary } from "~/composables/main/summary";
 import { useShortcutKeyMode } from "~/composables/user/shortcutKey";
-import { useCourseStore } from "~/store/course";
 import { cancelShortcut, registerShortcut } from "~/utils/keyboardShortcuts";
 
 const { shortcutKeys } = useShortcutKeyMode();
 const { playSound } = usePlaySound(shortcutKeys.value.sound);
 const { toggleGameMode } = useShowAnswer(shortcutKeys.value.answer);
-const { goToNextQuestion, goToPreviousQuestion } = usePrevAndNextQuestion(
-  shortcutKeys.value.previous,
-  shortcutKeys.value.skip
-);
-
-const { showQuestion } = useGameMode();
-const { showSummary } = useSummary();
-const courseStore = useCourseStore();
 
 const toggleTipText = computed(() => {
   let text = "";
@@ -154,41 +127,6 @@ function useShowAnswer(key: string) {
   };
 }
 
-// 上一题/下一题
-function usePrevAndNextQuestion(prevKey: string, nextKey: string) {
-  handleShortcut();
-
-  function goToNextQuestion() {
-    if (courseStore.isAllDone()) {
-      showSummary();
-      return;
-    }
-    courseStore.toNextStatement();
-    showQuestion();
-  }
-
-  function goToPreviousQuestion() {
-    courseStore.toPreviousStatement();
-    showQuestion();
-  }
-
-  function handleShortcut() {
-    onMounted(() => {
-      registerShortcut(nextKey, goToNextQuestion);
-      registerShortcut(prevKey, goToPreviousQuestion);
-    });
-
-    onUnmounted(() => {
-      cancelShortcut(prevKey, goToNextQuestion);
-      cancelShortcut(nextKey, goToPreviousQuestion);
-    });
-  }
-
-  return {
-    goToNextQuestion,
-    goToPreviousQuestion,
-  };
-}
 </script>
 
 <style scoped>
